+++ weblate/addons/tests.py
@@ -21,6 +21,7 @@ from django.test import TestCase
 from django.test.utils import override_settings
 from django.urls import reverse
 from django.utils import timezone
+from standardwebhooks.webhooks import Webhook, WebhookVerificationError
 
 from weblate.lang.models import Language
 from weblate.trans.actions import ActionEvents
@@ -68,7 +69,7 @@ from .properties import PropertiesSortAddon
 from .removal import RemoveComments, RemoveSuggestions
 from .resx import ResxUpdateAddon
 from .tasks import cleanup_addon_activity_log, daily_addons
-from .webhooks import StandardWebhooksUtils, WebhookAddon, WebhookVerificationError
+from .webhooks import WebhookAddon
 from .xml import XMLCustomizeAddon
 from .yaml import YAMLCustomizeAddon
 
@@ -1649,6 +1650,7 @@ class WebhookAddonsTest(ViewTestCase):
     addon_configuration: ClassVar[dict] = {
         "webhook_url": "https://example.com/webhooks",
         "events": [],
+        "secret": "XXXXXXXX",
     }
 
     def setUp(self) -> None:
@@ -1850,13 +1852,13 @@ class WebhookAddonsTest(ViewTestCase):
         self.do_translation_added_test(response_code=200)
 
         wh_request = responses.calls[0].request
-        wh_utils = StandardWebhooksUtils("secret-string")
+        wh_utils = Webhook("secret-string")
 
         # valid request
         wh_utils.verify(wh_request.body, wh_request.headers)
 
         # valid request with bytes
-        StandardWebhooksUtils(base64.b64decode("secret-string")).verify(
+        Webhook(base64.b64decode("secret-string")).verify(
             wh_request.body, wh_request.headers
         )
 
@@ -1887,7 +1889,7 @@ class WebhookAddonsTest(ViewTestCase):
 
         #  "Invalid secret"
         with self.assertRaises(WebhookVerificationError):
-            StandardWebhooksUtils("xxxx-xxxx-xxxx").verify(wh_request.body, wh_headers)
+            Webhook("xxxx-xxxx-xxxx").verify(wh_request.body, wh_headers)
 
         #  "Invalid format of timestamp"
         with self.assertRaises(WebhookVerificationError):
@@ -1915,7 +1917,7 @@ class WebhookAddonsTest(ViewTestCase):
         """Test WebhooksAddonForm."""
         self.user.is_superuser = True
         self.user.save()
-        # Missing params
+        # Missing url param
         response = self.client.post(
             reverse("addons", kwargs=self.kw_component),
             {
@@ -1926,26 +1928,18 @@ class WebhookAddonsTest(ViewTestCase):
         )
         self.assertNotContains(response, "Installed 1 add-on")
 
-        # empty secret
+        # empty secret not allowed
         response = self.client.post(
             reverse("addons", kwargs=self.kw_component),
             {
                 "name": "weblate.webhook.webhook",
                 "form": "1",
                 "webhook_url": "https://example.com/webhooks",
+                "secret": "",
             },
             follow=True,
         )
-        self.assertContains(response, "Installed 1 add-on")
-
-        # delete addon
-        addon_id = Addon.objects.get(component=self.component).id
-        response = self.client.post(
-            reverse("addon-detail", kwargs={"pk": addon_id}),
-            {"delete": "weblate.webhook.webhook"},
-            follow=True,
-        )
-        self.assertContains(response, "No add-ons currently installed")
+        self.assertNotContains(response, "Installed 1 add-on")
 
         # invalid secret
         response = self.client.post(
@@ -1974,3 +1968,12 @@ class WebhookAddonsTest(ViewTestCase):
             follow=True,
         )
         self.assertContains(response, "Installed 1 add-on")
+
+        # delete addon
+        addon_id = Addon.objects.get(component=self.component).id
+        response = self.client.post(
+            reverse("addon-detail", kwargs={"pk": addon_id}),
+            {"delete": "weblate.webhook.webhook"},
+            follow=True,
+        )
+        self.assertContains(response, "No add-ons currently installed")