+++ weblate/addons/webhooks.py
@@ -4,12 +4,9 @@
 
 from __future__ import annotations
 
-import base64
 import hashlib
 import hmac
 import json
-from datetime import UTC, datetime, timedelta
-from math import floor
 from typing import TYPE_CHECKING
 
 import jsonschema.exceptions
@@ -18,6 +15,7 @@ from django.template.loader import render_to_string
 from django.utils import timezone as dj_timezone
 from django.utils.translation import gettext_lazy
 from drf_spectacular.utils import OpenApiResponse, OpenApiWebhook, extend_schema
+from standardwebhooks.webhooks import Webhook
 from weblate_schemas import load_schema, validate_schema
 
 from weblate.addons.base import ChangeBaseAddon
@@ -32,10 +30,6 @@ if TYPE_CHECKING:
 
 def hmac_data(key: bytes, data: bytes) -> bytes:
     return hmac.new(key, data, hashlib.sha256).digest()
-
-
-class WebhookVerificationError(Exception):
-    """Exception raised when the payload cannot be validated."""
 
 
 class MessageNotDeliveredError(Exception):
@@ -124,7 +118,7 @@ class WebhookAddon(ChangeBaseAddon):
         self, change: Change, payload: dict[str, int | str | list[str]]
     ) -> dict[str, str]:
         """Build headers following Standard Webhooks specifications."""
-        wh = StandardWebhooksUtils(self.instance.configuration.get("secret", ""))
+        wh = Webhook(self.instance.configuration.get("secret"))
         webhook_id = change.get_uuid().hex
         attempt_time = dj_timezone.now()
         return {
@@ -138,84 +132,6 @@ class WebhookAddon(ChangeBaseAddon):
             "addons/webhook_log.html",
             {"activity": activity, "details": activity.details["result"]},
         )
-
-
-class StandardWebhooksUtils:
-    """Class providing utils for Standard Webhooks specification."""
-
-    _SECRET_PREFIX: str = "whsec_"  # noqa: S105
-    _whsecret: bytes
-    SIG_VERSION: str = "v1"
-
-    def __init__(self, whsecret: str | bytes):
-        if isinstance(whsecret, str):
-            whsecret = whsecret.removeprefix(self._SECRET_PREFIX)
-            self._whsecret = base64.b64decode(whsecret)
-
-        elif isinstance(whsecret, bytes):
-            self._whsecret = whsecret
-
-    def verify(self, data: bytes | str, headers: dict[str, str]):
-        """
-        Verify that the data has not been tempered.
-
-        :param data: The data to verify
-        :param headers: The headers to verify
-
-        :raises weblate.addons.webhooks.WebhookVerificationError: If the request
-            is not verified
-
-        :return: The data as a JSON object if the request is verified
-        """
-        data = data if isinstance(data, str) else data.decode()
-        headers = {k.lower(): v for (k, v) in headers.items()}
-        msg_id = headers.get("webhook-id")
-        msg_signature = headers.get("webhook-signature")
-        msg_timestamp = headers.get("webhook-timestamp")
-        if not (msg_id and msg_timestamp and msg_signature):
-            msg = "Missing required headers"
-            raise WebhookVerificationError(msg)
-
-        timestamp = self.__verify_timestamp(msg_timestamp)
-
-        expected_sig = base64.b64decode(
-            self.sign(msg_id=msg_id, timestamp=timestamp, data=data).split(",")[1]
-        )
-        passed_sigs = msg_signature.split(" ")
-        for versioned_sig in passed_sigs:
-            (version, signature) = versioned_sig.split(",")
-            if version != self.SIG_VERSION:
-                continue
-            sig_bytes = base64.b64decode(signature)
-            if hmac.compare_digest(expected_sig, sig_bytes):
-                return json.loads(data)
-        msg = "No matching signature found"
-        raise WebhookVerificationError(msg)
-
-    def sign(self, msg_id: str, timestamp: datetime, data: str) -> str:
-        """Generate a unique signature for payload."""
-        timestamp_str = str(floor(timestamp.replace(tzinfo=UTC).timestamp()))
-        to_sign = f"{msg_id}.{timestamp_str}.{data}".encode()
-        signature = hmac_data(self._whsecret, to_sign)
-        return f"{self.SIG_VERSION},{base64.b64encode(signature).decode('utf-8')}"
-
-    def __verify_timestamp(self, timestamp_header: str) -> datetime:
-        """Verify if timestamp from header is valid."""
-        webhook_tolerance = timedelta(minutes=5)
-        now = datetime.now(tz=UTC)
-        try:
-            timestamp = datetime.fromtimestamp(float(timestamp_header), tz=UTC)
-        except Exception as error:
-            msg = "Invalid Signature Headers"
-            raise WebhookVerificationError(msg) from error
-
-        if timestamp < (now - webhook_tolerance):
-            msg = "Message timestamp too old"
-            raise WebhookVerificationError(msg)
-        if timestamp > (now + webhook_tolerance):
-            msg = "Message timestamp too new"
-            raise WebhookVerificationError(msg)
-        return timestamp
 
 
 change_event_webhook = OpenApiWebhook(